refactoring: avatar component  - component  - docs
