--- docs/components/avatar/examples/AvatarAlternativeTextExample.vue
@@ -1,8 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Avatar status="online" alt="Alternative text" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"/>
-  </div>
-</template>
-<script setup>
-import { Avatar } from '../../../../src/index'
-</script>
