--- docs/components/avatar/examples/AvatarBorderedExample.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Avatar status="online" bordered img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-    <Avatar status="online" bordered rounded img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-  </div>
-</template>
-<script setup>
-import { Avatar } from '../../../../src/index'
-</script>
