--- docs/components/avatar/examples/AvatarDotIndicatorExample.vue
@@ -1,11 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Avatar img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" status="online" />
-    <Avatar img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" status="busy" />
-    <Avatar img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" status="away" />
-    <Avatar img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" status="offline" />
-  </div>
-</template>
-<script setup>
-import { Avatar } from '../../../../src/index'
-</script>
