--- docs/components/avatar/examples/AvatarExample.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Avatar status="online" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-    <Avatar status="online" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" rounded />
-  </div>
-</template>
-<script setup>
-import { Avatar } from '../../../../src/index'
-</script>
