--- docs/components/avatar/examples/AvatarPlaceholderExample.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Avatar class="mr-2.5" />
-    <Avatar rounded />
-  </div>
-</template>
-<script setup>
-import { Avatar } from '../../../../src/index'
-</script>
