--- docs/components/avatar/examples/AvatarPlaceholderInitialsExample.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Avatar initials="JD" class="mr-2.5" />
-    <Avatar initials="JD" rounded />
-  </div>
-</template>
-<script setup>
-import { Avatar } from '../../../../src/index'
-</script>
