--- docs/components/avatar/examples/AvatarSizeExample.vue
@@ -1,14 +1,0 @@
-<template>
-  <div class="vp-raw">
-    <div class="flex items-center">
-      <Avatar size="xs" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-      <Avatar size="sm" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-      <Avatar size="md" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-      <Avatar size="lg" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-      <Avatar size="xl" img="https://flowbite.com/docs/images/people/profile-picture-5.jpg" class="mr-2.5" />
-    </div>
-  </div>
-</template>
-<script setup>
-import { Avatar } from '../../../../src/index'
-</script>
