+++ docs/components/avatar/examples/FwbAvatarExample.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="vp-raw flex justify-center space-x-4">
+    <fwb-avatar
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+    />
+    <fwb-avatar
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+      rounded
+    />
+  </div>
+</template>
+
+<script setup>
+import { FwbAvatar } from '../../../../src/index'
+</script>
