+++ docs/components/avatar/examples/FwbAvatarExampleAlt.vue
@@ -1,0 +1,12 @@
+<template>
+  <div class="vp-raw flex justify-center">
+    <fwb-avatar
+      alt="Alternative text"
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+    />
+  </div>
+</template>
+
+<script setup>
+import { FwbAvatar } from '../../../../src/index'
+</script>
