+++ docs/components/avatar/examples/FwbAvatarExampleInitials.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="vp-raw flex justify-center space-x-4">
+    <fwb-avatar initials="JD" />
+    <fwb-avatar
+      initials="JD"
+      rounded
+    />
+  </div>
+</template>
+
+<script setup>
+import { FwbAvatar } from '../../../../src/index'
+</script>
