+++ docs/components/avatar/examples/FwbAvatarExamplePlaceholder.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="vp-raw flex justify-center space-x-4">
+    <fwb-avatar />
+    <fwb-avatar rounded />
+  </div>
+</template>
+
+<script setup>
+import { FwbAvatar } from '../../../../src/index'
+</script>
