+++ docs/components/avatar/examples/FwbAvatarExampleSize.vue
@@ -1,0 +1,28 @@
+<template>
+  <div class="vp-raw flex justify-center items-center space-x-4">
+    <fwb-avatar
+      size="xs"
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+    />
+    <fwb-avatar
+      size="sm"
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+    />
+    <fwb-avatar
+      size="md"
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+    />
+    <fwb-avatar
+      size="lg"
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+    />
+    <fwb-avatar
+      size="xl"
+      img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+    />
+  </div>
+</template>
+
+<script setup>
+import { FwbAvatar } from '../../../../src/index'
+</script>
